NEW FEATURE - Managers Portal
The basis of the portal is to set information in a managers only portal accessible by the signed in users job title within their MS 365 account.

Allow "managers" to set learning for employees in their store group and check on progress.

With the staff member, the learnings have been assigned they must acknowledge the information which has been returned to the manager.

No database has been added yet but the server receives the data being set and then arrays of json data can be served to the user.