Transmitting Quill data to server in JSON format.
With the addition to the quill editor in the notes section of this project, the end goal is to create a pdf on the server to be sent back to the user.

So far the delta values from quill are being sent to the server to be next inserted as values into makePDF and then the generated pdf sent back to the user.