Quill Text editor & Puppeteer for PDF generation implemented.
The text editor has been updated with quill, this allows for different font sizes and elements like lists, the data in this editor can be sent over http as a delta json array of data. On the receiving end it is converted to a html string and placed in a template. Handlebars is then used to compile the data into a object.

Puppeteer then takes this data and generates a pdf.

:)